requirements.txt
openai==1.61.1
python-dotenv==1.0.1
langgraph==0.2.69
langgraph-checkpoint-sqlite==2.0.3
langchain-openai==0.3.3
tavily-python==0.5.1
langchain-community==0.3.16
tavily-python==0.5.1
pandas==2.2.3
streamlit==1.42.0